Advanced AI Players
I've been working on a FiveM script that introduces advanced AI players into your server. Basically, the actual AI part isnt done and will take months, so I'm posting this to give you something to play around with. So this version (1.0.2) pretty much will do checks to see if the server has any players that need to be revived, initiates the player revival system and displays messaged on the right side of the screen to keep the player updated while they hold on for dear life. There are several AI Players posted strategically, so that no matter where you are, there is always an AI close-by and ready to assist.